requirements.txt
fastapi
uvicorn
httpx
python-multipart
pydantic
jinja2
pyyaml